apps/gateway/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-extensions==3.2.3
django-redis==5.4.0
requests==2.9.7
redis==5.0.1
celery==5.3.4
prometheus-client==0.19.0
structlog==23.2.0
python-json-logger==2.0.7
pydantic==2.5.0
httpx==0.25.2
python-dotenv==1.0.0
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
faker==20.1.0
